third_party/blink/web_tests/platform/generic/fullscreen/full-screen-iframe-ua-style-expected.txt

TEST(getComputedStyle(document.documentElement).overflow == 'visible') OK
EVENT(webkitfullscreenchange)
TEST(getComputedStyle(document.documentElement).overflow == 'hidden') OK
END OF TEST
